requirements.txt
python-pptx
streamlit
google-generativeai
